Move SQS schemas for reuse
