+++ worker/schemas/sqs.py
@@ -1,0 +1,27 @@
+import json
+from dataclasses import dataclass
+from typing import List, Union
+
+
+@dataclass
+class SQSMessageRecord:
+    messageId: str
+    receiptHandle: str
+    body: Union[dict, str]
+    attributes: dict
+    messageAttributes: dict
+    md5OfBody: str
+    eventSource: str
+    eventSourceARN: str
+    awsRegion: str
+
+    def __post_init__(self):
+        self.body = json.loads(self.body) if type(self.body) == str else self.body
+
+
+@dataclass
+class SQSMessage:
+    Records: Union[List[SQSMessageRecord], dict]
+
+    def __post_init__(self):
+        self.Records = [SQSMessageRecord(**x) for x in self.Records]
